+++ f427 dipan jieusan+yaokong/Hardware/can_user.h
@@ -13,7 +13,7 @@
 extern motor_recieve  motor_recieve_dipan3508[4];;
 
 void CAN_cmd_current_3508motor(int16_t motor1, int16_t motor2, int16_t motor3, int16_t motor4);
-void CAN_cmd_speed_3508motor(float target[4], motor_recieve motor_recieve_info[4]);
+void CAN_cmd_speed_3508motor(int16_t target[4], motor_recieve motor_recieve_info[4]);
 void can_filter_init(void);
 void HAL_CAN_RxFifo0MsgPendingCallback(CAN_HandleTypeDef *hcan);
